Rename `PanGalacticCommercialEnterprise` to `PanCosmicCommercialEnterprise`
