+++ libs/sr/assets/src/gen/relic/PanCosmicCommercialEnterprise/rope.png
