+++ libs/sr/assets/src/gen/relic/PanCosmicCommercialEnterprise/sphere.png
